backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
sqlalchemy==2.0.36
alembic==1.14.0
psycopg2-binary==2.9.10
python-dotenv==1.0.1
pydantic==2.10.3
pytest==8.3.4
pytest-cov==6.0.0
httpx==0.28.1
ruff==0.8.4
